requirements.txt
amqp==5.1.1
asgiref==3.6.0
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
billiard==3.6.4.0
black==23.1.0
celery==5.2.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cron-descriptor==1.2.35
cryptography==40.0.1
daphne==4.0.0
dj-database-url==1.2.0
Django==4.2
django-celery-beat==2.5.0
django-configurations==2.4.1
django-environ==0.10.0
django-filter==23.1
django-nose==1.4.7
django-redis==5.2.0
django-timezone-field==5.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
environ==1.0
Faker==18.4.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kombu==5.2.4
msgpack==1.0.5
mypy-extensions==1.0.0
nose==1.3.7
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
prompt-toolkit==3.0.38
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.1.1
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.2
redis==4.5.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2023.2
vine==5.0.0
wcwidth==0.2.6
zope.interface==6.0
